Create Article - Create tests; - Fix truncate function; - Improved factories to generate random data with each test; - Tests code refactor;

+++ __tests__/util/truncate.js
@@ -1,11 +1,14 @@
 import database from '../../src/database';
 
-export default function truncate() {
-  return Promise.all(
+(async () => await database.connection.query('PRAGMA foreign_keys = OFF'))();
+
+export default async function truncate() {
+  return await Promise.all(
     Object.keys(database.connection.models).map(key => {
+      if (['sequelize', 'Sequelize'].includes(key)) return null;
       return database.connection.models[key].destroy({
-        truncate: true,
         force: true,
+        truncate: true,
       });
     })
   );
